src/main/java/com/Aashish/online_platform/project/controller/MainController.java
package com.Aashish.online_platform.project.controller;

import java.util.List;

import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;

import com.Aashish.online_platform.project.dto.UserDto;
import com.Aashish.online_platform.project.service.UserService;

@Controller
public class MainController {

    private final UserService userService;

    public MainController(UserService userService) {
        this.userService = userService;
    }

    @GetMapping("/")
    public String home() {
        return "home"; 
    }

    @GetMapping("/home")
    public String homePage() {
        return "home"; 
    }


    @GetMapping("/users")
    public String listUsers(Model model) {
        List<UserDto> users = userService.findAllUsers();
        model.addAttribute("userRecords", users); 
        return "users/userRecords"; 
    }

    @DeleteMapping("/delete/{id}")
    public String deleteUser(@PathVariable Long id) {
        userService.deleteUser(id);
        return "redirect:/users"; 
    }
}
